Refactor: Separate Fonts & Songs gui to their own modules
